TSP: rename x and y to latitude and longitude + implement distance function as specified by TSPLIB

+++ drools-planner-examples/src/main/java/org/drools/planner/examples/tsp/domain/CityAssignment.java
@@ -115,7 +115,7 @@ public class CityAssignment extends AbstractPersistable implements Comparable<Ci
         return city.toString();
     }
 
-    public double getDistanceToNextCityAssignment() {
+    public int getDistanceToNextCityAssignment() {
         return city.getDistance(nextCityAssignment.getCity());
     }
